refactor: move BackendCommunicator to server package and separate from common Communicator
